fix: Resolve all 11 mypy type checking errors
Fixed type checking errors across 4 files:

- syllables.py: Added type ignore comments for untyped pronouncing library
- smog.py/_types.py: Changed SMOGResult.grade_level to float to support NaN
- hapax.py: Added type ignore comments for bnc_lookup/wordnet_lookup
- complex_words.py: Renamed variable to fix type inference in basic mode

All mypy checks now pass cleanly.

+++ pystylometry/_types.py
@@ -91,7 +91,7 @@ class SMOGResult:
     """Result from SMOG Index computation."""
 
     smog_index: float
-    grade_level: int
+    grade_level: float
     metadata: dict[str, Any]
